+++ src/main/java/dev/kwameopareasiedu/simpson/nodes/DecimalNode.java
@@ -1,0 +1,8 @@
+package dev.kwameopareasiedu.simpson.nodes;
+
+/** {@link DecimalNode} is a {@link Node} whose value is a double */
+public class DecimalNode extends Node<Double> {
+  public DecimalNode(Double value) {
+    super(Type.DECIMAL, value);
+  }
+}