+++ src/main/java/dev/kwameopareasiedu/simpson/nodes/IntegerNode.java
@@ -1,0 +1,8 @@
+package dev.kwameopareasiedu.simpson.nodes;
+
+/** {@link IntegerNode} is a {@link Node} whose value is an integer */
+public class IntegerNode extends Node<Integer> {
+  public IntegerNode(Integer value) {
+    super(Type.INTEGER, value);
+  }
+}