+++ src/main/java/dev/kwameopareasiedu/simpson/nodes/StringNode.java
@@ -1,0 +1,8 @@
+package dev.kwameopareasiedu.simpson.nodes;
+
+/** {@link StringNode} is a {@link Node} whose value is a string */
+public class StringNode extends Node<String> {
+  public StringNode(String value) {
+    super(Type.STRING, value);
+  }
+}